got login with pbkdf2 and counter mode working

+++ primitives/pbkdf2.h
@@ -7,14 +7,15 @@
 typedef std::string (*prf)(std::string, std::string);
 prf current_prf;
 
-// c should be at least 600,000 according to owasp
+// c is the number of iterations; should be at least 600,000 according to owasp
 // dkLen is the bitlength of the derived key
+// salt should be at least 8 chars long, and randomly generated 
 std::string pbkdf2(prf prf_param, std::string password, std::string salt, int c, int dkLen) {
     current_prf = prf_param;
     int hLen = current_prf("test key", "test message").size() * 4;
     
     std::string dk = "";
-    for (uint32_t i = 1; i <= dkLen/hLen; i++) {
+    for (int i = 1; i <= dkLen/hLen; i++) {
         std::string augmented_salt = salt;
         augmented_salt += static_cast<unsigned char>(i >> 24);
         augmented_salt += static_cast<unsigned char>(i >> 16);
@@ -25,7 +26,7 @@ std::string pbkdf2(prf prf_param, std::string password, std::string salt, int c,
         for (int j = 2; j <= c; j++) {
             std::string next = hex_to_utf8(current_prf(password, prev)); 
             prev = next;
-            for (int k = 0; k < T.size(); k++) {
+            for (unsigned long k = 0; k < T.size(); k++) {
                 T[k] = static_cast<unsigned char>(T[k] ^ next[k]); 
             }
         }
